django_mongo/requirements.txt
apipkg==1.4
appdirs==1.4.3
appnope==0.1.0
astroid==1.5.3
attrs==17.3.0
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.11
Django==2.0.1
django-autofixture==0.12.1
django-rest-swagger==2.1.2
djangorestframework==3.7.7
drf-nested-routers==0.90.0
execnet==1.5.0
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
openapi-codec==1.3.1
packaging==16.8
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.14
psycopg2==2.7.3.2
ptyprocess==0.5.1
py==1.5.2
Pygments==2.2.0
pylint==1.7.2
pymongo==3.6.0
pyparsing==2.2.0
pytest==3.3.1
pytest-cache==1.0
pytest-pep8==1.0.6
pytz==2017.3
requests==2.18.1
rethinkdb==2.3.0.post6
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.21.1
wcwidth==0.1.7
wrapt==1.10.10
